test: consolidate python tests into test_pgqrs.py
